refactor: update get partial request
