+++ src/message/models/messageManager.ts
@@ -3,8 +3,8 @@ import { inject, injectable } from 'tsyringe';
 import { v4 as uuidv4 } from 'uuid';
 import type { components } from '@openapi';
 import { SERVICES } from '@common/constants';
+import { localMesssagesStore } from '../../common/mocks';
 import { IQueryModel } from './../../common/interfaces';
-import { localMesssagesStore } from './../../common/payloads';
 
 export type ILogObject = components['schemas']['ILogObject'];
